requirements.txt
Keras==2.0.4
numpy==1.12.1
tensorflow-gpu==1.1.0
h5py
